refactor(output): isolate savings policy
